+++ src/sessions/v2/sessions-v2.test.ts
@@ -103,6 +103,7 @@ const generateTestData = async () => {
       paymentIntegrationSubscriptionId: stripeEnvs.basicSubscriptionFourId,
       uuid: subscriptionUuid,
       email: 'tester@testing.com',
+      owner: 'owner_12345',
       type: 'salable',
       status: 'ACTIVE',
       organisation: testUuids.organisationId,